# public/openapi.yaml
openapi: "3.0.3"
info:
  title: 智学AI - 开放API
  description: |
    智学AI教育平台开放API接口文档。提供AI辅导、LLM代理、题库管理、游戏化学习等接口。

    ## 快速开始
    1. **注册账号** - 在智学AI平台注册账号
    2. **获取API Key** - 在个人中心的「API Keys」页面创建虚拟API Key
    3. **调用API** - 使用 `X-API-Key` 请求头携带API Key发起请求

    ## 认证方式
    - **API Key** (`ApiKeyAuth`): 通过 `X-API-Key` 请求头或 `Authorization: Bearer <key>` 传递
    - **会话认证** (`BearerAuth`): 通过 NextAuth 的 session cookie 自动认证（仅限平台内调用）

    ## 错误响应
    所有接口在发生错误时都会返回统一格式：
    ```json
    { "error": "错误描述信息" }
    ```

    常见状态码：
    - `400` - 请求参数错误
    - `401` - 未认证 / API Key无效
    - `402` - 余额不足
    - `403` - 无权限
    - `404` - 资源不存在
    - `429` - 请求频率超限
    - `500` - 服务器内部错误

    ## 速率限制
    - 免费用户: 50 RPM
    - VIP用户: 200 RPM
    超出限制后返回 `429` 状态码。

  version: "1.0.0"
  contact:
    name: 智学AI开发团队
    email: dev@zixueai.com
    url: https://zixueai.com
servers:
  - url: "{baseUrl}"
    description: 智学AI API服务器
    variables:
      baseUrl:
        default: "https://api.zixueai.com"
        description: API服务器地址
        enum:
          - "https://api.zixueai.com"
          - "http://localhost:3000"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "从个人中心获取的API Key，格式: sk-edu-xxxxxx-xxxxxxxx"
    BearerAuth:
      type: http
      scheme: bearer
      description: "NextAuth Session Token（平台内自动携带）"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误描述信息
          example: "请先登录"
      required:
        - error
...
    Question:
      type: object
      properties:
        id:
          type: string
          example: "q_abc123"
        subject:
          type: string
          example: "数学"
        type:
          type: string
          enum: [choice, fill, essay]
          example: "choice"
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        content:
          type: string
          example: "在直角三角形ABC中..."
        options:
          type: array
          items:
            type: string
          example: ["A. 勾股定理", "B. 相似三角形", "C. 面积公式", "D. 三角函数"]
        answer:
          type: string
          example: "A"
        explanation:
          type: string
          example: "勾股定理是直角三角形的基本性质..."
        source:
          type: string
          example: "2024年中考真题"
        gradeLevel:
          type: string
          example: "初三"
        knowledgePoints:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              subject:
                type: string
        errorNoteCount:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
    QuestionListResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
          example: 1200
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
...
    DailyChallenge:
      type: object
      properties:
        id:
          type: string
          example: "dc_abc123"
        date:
          type: string
          format: date
          example: "2026-05-07"
        question:
          $ref: "#/components/schemas/Question"
        completed:
          type: boolean
          example: false
        score:
          type: integer
          description: 得分（0-100），完成才有
        createdAt:
          type: string
          format: date-time
    KnowledgeGraphResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
                example: "勾股定理"
              subject:
                type: string
                example: "数学"
              description:
                type: string
              gradeLevel:
                type: string
                example: "初二"
              parentId:
                type: string
              orderIndex:
                type: integer
              masteryLevel:
                type: integer
                description: 掌握度百分比 0-100
                example: 85
              practiceCount:
                type: integer
                example: 12
              lastPracticedAt:
                type: string
                format: date-time
        edges:
          type: array
          items:
            type: object
            properties:
              from:
                type: string
                description: 父节点ID
              to:
                type: string
                description: 子节点ID
              relation:
                type: string
                enum: [parent, prerequisite]
                description: 关系类型
...
    SafetyCheckRequest:
      type: object
      properties:
        content:
          type: string
          description: 待检测的文本内容
          example: "你好，请问这道题怎么做？"
          minLength: 1
        type:
          type: string
          enum: [input, output]
          default: "input"
          description: 检测类型：input 用户输入，output AI输出
      required:
        - content
    SafetyCheckResponse:
      type: object
      properties:
        safe:
          type: boolean
          example: true
        reason:
          type: string
          description: 不安全原因（safe=false时返回）
        severity:
          type: string
          enum: [low, medium, high, critical]
        action:
          type: string
          description: 建议的处理方式
        checkedAt:
          type: string
          format: date-time
...
paths:
...
  /api/questions/search:
    get:
      tags:
        - Questions
      summary: 搜索题目
      description: 按关键词搜索题目，支持内容、答案、解析等多字段匹配
      operationId: searchQuestions
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          description: 搜索关键词（必填）
          required: true
          schema:
            type: string
          example: "勾股定理"
        - name: subject
          in: query
          description: 学科筛选
          required: false
          schema:
            type: string
          example: "数学"
        - name: difficulty
          in: query
          description: 难度等级（1-5）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: gradeLevel
          in: query
          description: 年级筛选
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: 每页数量（最大100）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: 偏移量
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionListResponse"
        "400":
          description: 搜索关键词为空
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "请输入搜索关键词"

  # ==================== Error Notes ====================
...
  /api/game/daily-challenge:
    get:
      tags:
        - Game
      summary: 获取每日挑战
      description: |
        获取当天的每日挑战题目。如果当天还没有挑战记录，会自动创建新的挑战（随机抽取一道题目）。
      operationId: getDailyChallenge
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 每日挑战题目
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyChallenge"
              example:
                id: "dc_abc123"
                date: "2026-05-07"
                question:
                  id: "q_abc123"
                  subject: "数学"
                  type: "choice"
                  difficulty: 3
                  content: "对于二次函数 y=ax^2+bx+c，其对称轴方程是？"
                  options:
                    - "A. x=-b/2a"
                    - "B. x=b/2a"
                    - "C. x=-b/a"
                    - "D. x=b/a"
                  answer: "A"
                  explanation: "二次函数的对称轴公式为 x = -b/(2a)"
                  source: "人教版数学"
                  gradeLevel: "初三"
                  knowledgePoints:
                    - id: "kp_002"
                      name: "二次函数"
                      subject: "数学"
                  errorNoteCount: 3
                  createdAt: "2026-04-15T08:00:00Z"
                completed: false
                createdAt: "2026-05-07T00:01:00Z"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 暂无可用的题目
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "暂无可用的题目"

  # ==================== Knowledge Graph ====================
  /api/knowledge-graph:
    get:
      tags:
        - Knowledge
      summary: 获取知识图谱
      description: |
        获取指定学科的知识图谱数据，包含用户对每个知识点的掌握度。
        知识点以节点形式组织，通过边表示父子关系和前置关系。
        默认学科为"数学"。
      operationId: getKnowledgeGraph
      security:
        - BearerAuth: []
      parameters:
        - name: subject
          in: query
          description: 学科名称
          required: false
          schema:
            type: string
            default: "数学"
          example: "数学"
      responses:
        "200":
          description: 知识图谱数据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraphResponse"
              example:
                nodes:
                  - id: "kp_001"
                    name: "勾股定理"
                    subject: "数学"
                    description: "直角三角形两直角边的平方和等于斜边的平方"
                    gradeLevel: "初二"
                    parentId: null
                    orderIndex: 1
                    masteryLevel: 85
                    practiceCount: 12
                    lastPracticedAt: "2026-05-07T09:30:00Z"
                  - id: "kp_002"
                    name: "二次函数"
                    subject: "数学"
                    description: "形如y=ax^2+bx+c的函数"
                    gradeLevel: "初三"
                    parentId: null
                    orderIndex: 2
                    masteryLevel: 60
                    practiceCount: 5
                    lastPracticedAt: "2026-05-06T14:00:00Z"
                edges:
                  - from: "kp_001"
                    to: "kp_003"
                    relation: "prerequisite"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ==================== User ====================
...
  /api/safety/check:
    post:
      tags:
        - Safety
      summary: 内容安全检测
      description: |
        对用户输入或AI输出进行安全检测。检测不安全的文本内容，
        高风险事件会被记录到安全日志中。
      operationId: safetyCheck
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SafetyCheckRequest"
            examples:
              inputCheck:
                summary: 检测用户输入
                value:
                  content: "你好，请问这道题怎么做？"
                  type: "input"
              outputCheck:
                summary: 检测AI输出
                value:
                  content: "这道题的解法是..."
                  type: "output"
      responses:
        "200":
          description: 安全检测结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyCheckResponse"
              examples:
                safe:
                  summary: 内容安全
                  value:
                    safe: true
                    checkedAt: "2026-05-07T10:00:00Z"
                unsafe:
                  summary: 内容不安全
                  value:
                    safe: false
                    reason: "包含不当言论"
                    severity: "high"
                    action: "block"
                    checkedAt: "2026-05-07T10:00:00Z"
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "请提供待检测内容"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "请先登录"
        "402":
          description: 余额不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "余额不足，请充值后重试"
...